// apps/web/fsd/shared/lib/lightbox/types.ts
export type Slide = SlideTypes[SlideTypeKey];

/** Supported slide types */
export interface SlideTypes {
  /** image slide type */
  image: SlideImage;
}

/** Slide type key */
export type SlideTypeKey = keyof SlideTypes;

/** Generic slide */
export interface GenericSlide {
  type?: SlideTypeKey;
}

/** Image slide properties */
export interface SlideImage extends GenericSlide {
  /** image slide type */
  type?: 'image';
  /** image URL */
  src: string;
  /** image 'alt' attribute */
  alt?: string;
  /** image width in pixels */
  width?: number;
  /** image height in pixels */
  height?: number;
  /** `object-fit` setting */
  imageFit?: ImageFit;
  /** alternative images to be passed to the 'srcSet' */
  srcSet?: ImageSource[];
}

/** Image source */
export interface ImageSource {
  /** image URL */
  src: string;
  /** image width in pixels */
  width: number;
  /** image height in pixels */
  height: number;
}

/** Image fit setting */
export type ImageFit = 'contain' | 'cover';
// ...
export type ContainerRect = {
  width: number;
  height: number;
};
// ...
export interface CarouselSettings {
  /** if `true`, the lightbox carousel doesn't wrap around */
  finite: boolean;
  /** the lightbox preloads (2 * preload + 1) slides */
  preload: number;
  /** padding around each slide (e.g., "16px", "10%" or 0) */
  padding: LengthOrPercentage;
  /** spacing between slides (e.g., "100px", "50%" or 0) */
  spacing: LengthOrPercentage;
  /** `object-fit` setting for image slides */
  imageFit: ImageFit;
  /** custom image attributes */
  imageProps: Omit<
    React.ImgHTMLAttributes<HTMLImageElement>,
    'src' | 'alt' | 'sizes' | 'srcSet' | 'onLoad' | 'onError' | 'onClick'
  >;
}

export type LengthOrPercentage = `${number}px` | `${number}%` | number;
// ...
export type RenderFunction<T = void> = [T] extends [void]
  ? () => React.ReactNode
  : (props: T) => React.ReactNode;

/** `render.slide` render function props */
export interface RenderSlideProps<S extends Slide = Slide> {
  /** slide */
  slide: S;
  /** slide offset (`0` - current slide, `1` - next slide, `-1` - previous slide, etc.) */
  offset: number;
  /** container rect */
  rect: ContainerRect;
}

/** `render.slideHeader` render function props */
export interface RenderSlideHeaderProps {
  slide: Slide;
}

/** `render.slideFooter` render function props */
export interface RenderSlideFooterProps {
  slide: Slide;
}

/** `render.slideContainer` render function props */
export interface RenderSlideContainerProps extends React.PropsWithChildren {
  slide: Slide;
}

/** Custom render functions. */
export interface Render {
  /** render custom slide type, or override the default image slide */
  slide?: RenderFunction<RenderSlideProps>;
  /** render custom slide header (use absolute positioning) */
  slideHeader?: RenderFunction<RenderSlideHeaderProps>;
  /** render custom slide footer (use absolute positioning) */
  slideFooter?: RenderFunction<RenderSlideFooterProps>;
  /** render custom slide container */
  slideContainer?: RenderFunction<RenderSlideContainerProps>;
  /** render custom controls or additional elements in the lightbox (use absolute positioning) */
  controls?: RenderFunction;
  /** render custom Prev icon */
  iconPrev?: RenderFunction;
  /** render custom Next icon */
  iconNext?: RenderFunction;
  /** render custom Close icon */
  iconClose?: RenderFunction;
  /** render custom Loading icon */
  iconLoading?: RenderFunction;
  /** render custom Error icon */
  iconError?: RenderFunction;
  /** render custom Prev button */
  buttonPrev?: RenderFunction;
  /** render custom Next button */
  buttonNext?: RenderFunction;
  /** render custom Close button */
  buttonClose?: RenderFunction;
}